cpu_input_boost: Allow configuration of LP cluster's minimum idle frequency
Subsequent to:
7d8ef32 ("cpu_input_boost: Remove frame boosting entirely [1/2]")
2a875c8 ("drm/atomic: Remove frame boosting entirely [2/2]")

In lieu of maintaining a seperate boost level for rendering, we
will be globally raising the active minimum frequency for LP
cluster to 748MHz. In order to mitigate any negative impact this
may have on idle drain, allow for secondary configuration of LP's
screen-off minimum.

The way this is done currently is a bit of a hack, and should be
feasible to pull off without a dedicated get_idle_freq function.
However, for testing purposes we'll start with this explicit
implementation and optimize the code further down the line if
this works out.

Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>