[FIXUP] iommu: Update driver rewrite [7fa54ba / 7954d48]
This amends the following commit:
7fa54ba ("iommu: msm: Rewrite to improve clarity and performance")

This scope of this driver's lock usage is extremely wide, leading to
excessively long lock hold times. Other issues aside, this is the most
significant contributor to poor performance in this driver.

Fix everything by greatly reducing the scope of the locks used and by
using atomic reader/writer locks. The entire driver's code is refactored
and improved for better cleanliness and clarity.

This amends the following commit:
7954d48 ("iommu: msm: Fix race due to mapping and unmapping simultaneously")

A map can occur at the same time as an unmap, which can cause both the
map and unmap to fail in horrendous ways. Guard against this race by not
allowing mapping and unmapping for the same iommu meta to occur
simultaneously.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
[@0ctobot: Bring inline with kerneltoast/android_kernel_google_bluecross@82b0e20,
kerneltoast/android_kernel_google_wahoo@3149932, kerneltoast/android_kernel_google_wahoo@03e0313]
Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>