Revert "cpu_input_boost: Implement scalable Dynamic SchedTune Boosting"
The complexity of this is no longer necessary as we will be
increasing our base top-app SchedTune boost to 5. Additionally,
the inefficient means by which multiple boost slots were being
managed with this implementation may have been contributing to
spikes in total RT/CPU duration while rendering. Simplifying
Dynamic SchedTune Boosting in this manner will assist in
confirming that hypothesis.

This reverts commit 1194dd64de0077808182d83a71cbec5505d01d1c.

Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>